+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@scalefast/builder.io",
-  "version": "0.0.54",
+  "version": "0.0.55",
   "description": "",
   "entry": "plugin",
   "output": "plugin.system.js",
