+++ src/components/map/map-form.component.tsx
@@ -16,52 +16,9 @@ import {
 import { useEffect, useReducer, useState } from 'react';
 import { Clear as ClearIcon } from '@material-ui/icons';
 import { Add as AddIcon } from '@material-ui/icons';
-import { CustomApplicationContext, CustomMapOptions, CustomTypes, Model } from '../../models';
+import { CustomApplicationContext, CustomMapFormProps, CustomMapOptions, CustomTypes, Model } from '../../models';
 import { Column, getModels, Row } from '../../utils';
-
-interface CustomMapFormProps {
-  opt: CustomMapOptions;
-  onChange(action: string, index: number, newValue?: CustomMapOptions): void;
-  index: number;
-  context: CustomApplicationContext,
-}
-
-const actionMap: any = {
-  set_key: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    key: action.key,
-  }),
-  set_required: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    required: action.required,
-  }),
-  set_helper_text: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    helperText: action.helperText,
-  }),
-  set_type: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    type: action.t,
-  }),
-  set_values: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    values: action.values,
-  }),
-  set_tag: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    values: [...state.values ?? [], action.tag]
-  }),
-  set_extra_options: (state: CustomMapOptions, action: any) => ({
-    ...state,
-    extraOptions: action.extraOptions
-  })
-};
-
-const reducer = (state: CustomMapOptions, action: any) => {
-  const handler = actionMap[action.type] ?? null;
-
-  return handler ? handler(state, action) : state;
-};
+import { mapFormReducer } from './map-form-reducer.component';
 
 export const CustomMapForm = (props: CustomMapFormProps) => {
   const initialState = {
@@ -76,7 +33,7 @@ export const CustomMapForm = (props: CustomMapFormProps) => {
     initialState.extraOptions = props.opt.extraOptions;
   }
 
-  const [state, dispatch] = useReducer(reducer, initialState);
+  const [state, dispatch] = useReducer(mapFormReducer, initialState);
   const [tag, setTag] = useState<string>('');
   const [id] = useState<string>(`ìnput-component-prop-${props.index}`);
   const models: Model[] = getModels();
@@ -176,6 +133,7 @@ export const CustomMapForm = (props: CustomMapFormProps) => {
               ></TextField>
             </FormControl>
           </div>
+
           <div>
             <InputLabel css={{ display: 'block', paddingBottom: 10, fontSize: 14, fontWeight: 500 }} id={id + '-label'}>
               Type
@@ -201,7 +159,7 @@ export const CustomMapForm = (props: CustomMapFormProps) => {
                   ></TextField>
                 </FormControl>
               </div>
-              {state.type == 'select' && state.values && (
+              {state.values && (
                 <div>
                   {state.values.map((t: any, index: number) => (
                     <Chip
@@ -230,6 +188,7 @@ export const CustomMapForm = (props: CustomMapFormProps) => {
               </div>
             </Column>
           )}
+
         </Row>
 
         <Row css={{ marginTop: 10 }}>
